docs

+++ pax-core/src/expressions.rs
@@ -55,7 +55,7 @@ impl<T: Default + Clone> PropertyInstance<T> for PropertyExpression<T> {
     }
 
     //TODO: when trait fields land, DRY this implementation vs. other <T: PropertyInstance> implementations
-    fn ease_to(&mut self, new_value: T, duration_frames: usize, curve: EasingCurve) {
+    fn ease_to(&mut self, new_value: T, duration_frames: u64, curve: EasingCurve) {
         self.transition_manager.value = Some(self.get().clone());
         &self.transition_manager.queue.clear();
         &self.transition_manager.queue.push_back(TransitionQueueEntry {
@@ -67,7 +67,7 @@ impl<T: Default + Clone> PropertyInstance<T> for PropertyExpression<T> {
         });
     }
 
-    fn ease_to_later(&mut self, new_value: T, duration_frames: usize, curve: EasingCurve) {
+    fn ease_to_later(&mut self, new_value: T, duration_frames: u64, curve: EasingCurve) {
         if let None = self.transition_manager.value {
             //handle case where transition queue is empty -- a None value gets skipped, so populate it with Some
             self.transition_manager.value = Some(self.get().clone());
